add: navigation with bottom sheet mod: migrate to multiple screen with navigation
