+++ app/src/main/java/com/joeloewi/jumpkking/ui/navigation/JumpKkingNavigation.kt
@@ -1,0 +1,9 @@
+package com.joeloewi.jumpkking.ui.navigation
+
+sealed class JumpKkingNavigation(
+    val route: String,
+) {
+    object Friends : JumpKkingNavigation(
+        route = "friends"
+    )
+}
